Initial integration couchbase
