+++ pyon/datastore/couchbase/datastore.py
@@ -1,215 +1,58 @@
 #!/usr/bin/env python
 
-__author__ = 'Thomas R. Lennan, Michael Meisinger'
-__license__ = 'Apache 2.0'
+__author__ = 'Michael Meisinger, Seman Said'
 
 from uuid import uuid4
+import simplejson as json
+import gevent
+from gevent import sleep
+import requests
 
-import hashlib
+import couchbase.client
+import couchbase.rest_client
+# Monkey patching
+couchbase.client.json = json
+couchbase.rest_client.json = json
 
 from couchbase.client import Couchbase, Bucket
 from couchbase.rest_client import RestHelper
 from couchbase.exception import BucketCreationException, BucketUnavailableException, MemcachedTimeoutException
-import requests
-from pyon.datastore.couchdb.views import get_couchdb_view_designs
+from couchbase.exception import MemcachedError
+from requests.exceptions import ConnectionError
 
-from couchdb.client import ViewResults, Row
-from couchdb.http import PreconditionFailed, ResourceConflict, ResourceNotFound
+
+from pyon.datastore.couchdb.views import get_couchdb_view_designs
 
 from pyon.core.bootstrap import get_obj_registry, CFG
 from pyon.core.exception import BadRequest, Conflict, NotFound, ServerError
 from pyon.core.object import IonObjectBase, IonObjectSerializer, IonObjectDeserializer
 from pyon.datastore.datastore import DataStore
-from pyon.ion.identifier import create_unique_association_id
+from pyon.datastore.couchbase.base_store import CouchbaseDataStore
 from pyon.ion.resource import CommonResourceLifeCycleSM
 from pyon.util.log import log
 from pyon.util.arg_check import validate_is_instance
-from pyon.util.containers import get_ion_ts
-from pyon.util.stats import StatsCounter
-import simplejson as json
-from couchbase.exception import MemcachedError
-from requests.exceptions import ConnectionError
-from gevent import sleep
-
-import inspect
 
 
-import couchbase.client
-import couchbase.rest_client
-couchbase.client.json = json
-couchbase.rest_client.json = json
-
-
-# Token for a most likely non-inclusive key range upper bound (end_key), for queries such as
-# prefix <= keys < upper bound: e.g. ['some','value'] <= keys < ['some','value', END_MARKER]
-# or "somestr" <= keys < "somestr"+END_MARKER for string prefix checking
-# Note: Use highest ASCII characters here, not 8bit
-#END_MARKER = "\x7f\x7f\x7f\x7f"
-END_MARKER = "ZZZZZZ"
-
-
-class CouchbasePyonDataStore(DataStore):
+class CouchbasePyonDataStore(CouchbaseDataStore):
     """
-    Data store implementation utilizing CouchDB to persist documents.
-    For API info, see: http://packages.python.org/CouchDB/client.html
+    Pyon specialization of Couchbase datastore.
+    This class adds IonObject handling to the underlying base datastore.
     """
-    _stats = StatsCounter()
-    def __init__(self, host=None, port=None, datastore_name='default', options="", profile=DataStore.DS_PROFILE.BASIC):
-        log.debug('__init__(host=%s, port=%s, datastore_name=%s, options=%s)', host, port, datastore_name, options)
-        self.host = host or CFG.server.couchbase.host
-        self.port = port or CFG.server.couchbase.port
-        self.datastore_name = datastore_name
-        self.__attachment_string = '__attachments'  # Attachment string name that will be used in Couchbase
-        try:
-            if CFG.server.couchbase.username and CFG.server.couchbase.password:
-                log.debug("Using username:password authentication to connect to datastore")
-        except AttributeError:
-            log.error("Couchbase username:password not configured correctly ")
-            raise BadRequest ("Couchbase username:password not configured correctly ")
-        except ConnectionError:
-            log.error("Couchbase error: could not be connected to Couchbase.")
-            raise BadRequest ("Couchbase error: could not be connected to Couchbase.")
+    def __init__(self, datastore_name=None, profile=None, config=None, **kwargs):
+        log.debug('__init__(datastore_name=%s, profile=%s, config=%s)', datastore_name, profile, config)
 
-        self.username = CFG.server.couchbase.username
-        self.password = CFG.server.couchbase.password
-        connection_str = '%s:%s' %(self.host, self.port)
-        log.info('Connecting to Couchbase server: username:%s password:%s  "http://%s:%s"  ' % (self.username, self.password, self.host, self.port))
+        super(CouchbasePyonDataStore, self).__init__(datastore_name=datastore_name,
+                                                 config=config or CFG.get_safe("server.couchdb"),
+                                                 profile=profile or DataStore.DS_PROFILE.BASIC,
+                                                 newlog=log)
 
-        self.server = Couchbase(connection_str, username=self.username, password=self.password)
-        self.profile = profile
-
-
-        # serializers
+        # IonObject Serializers
         self._io_serializer = IonObjectSerializer()
-        # TODO: Not nice to have this class depend on ION objects
         self._io_deserializer = IonObjectDeserializer(obj_registry=get_obj_registry())
-        self._datastore_cache = {}
-
-    def close(self):
-        log.debug("Closing connection to Couchbase")
-        ##TODO:  is there a way to close connection
-        '''
-        ds, _ = self._get_datastore()
-        del ds
-        del self.server
-        '''
-
-    def _get_datastore(self, datastore_name=None):
-        datastore_name = datastore_name or self.datastore_name
-
-        if datastore_name in self._datastore_cache:
-            return (self._datastore_cache[datastore_name], datastore_name)
-
-        try:
-            ds = self.server[datastore_name]
-            self._datastore_cache[datastore_name] = ds
-            return ds, datastore_name
-        except ResourceNotFound:
-            raise BadRequest("Datastore '%s' does not exist" % datastore_name)
-        except ValueError:
-            raise BadRequest("Datastore name '%s' invalid" % datastore_name)
-
-    def create_datastore(self, datastore_name="", create_indexes=True, profile=None):
-        datastore_name = datastore_name or self.datastore_name
-        profile = profile or self.profile
-        bucket_password = CFG.server.couchbase.bucket_password,
-        ram_quota_mb =   CFG.server.couchbase.bucket_ram_quota_samll_mb
-        log.info('Creating data store %s with profile=%s' % (datastore_name, profile))
-        if self.datastore_exists(datastore_name):
-            raise BadRequest("Data store with name %s already exists" % datastore_name)
-        self._create_bucket(name=datastore_name, sasl_password=bucket_password, ram_quota_mb=ram_quota_mb)
-        if create_indexes:
-            self._define_views(datastore_name, profile)
-
-    def _create_bucket(self, name, auth_type='sasl', bucket_type='couchbase', parallel_db_and_view_compaction='false',
-                       ram_quota_mb="128", replica_index='0', replica_number='1', sasl_password=None, flush_enabled=False, proxy_port=11211):
-        '''
-        If you set authType to "none", then you must specify a proxyPort number.
-        If you set authType to "sasl", then you may optionally provide a "saslPassword" parameter.
-           For Couchbase Sever 1.6.0, any SASL authentication-based access must go through a proxy at port 11211.
-        '''
-        payload = dict()
-        payload['name'] = name
-        payload['authType'] = auth_type
-        payload['bucketType'] = bucket_type
-        if flush_enabled:
-            payload['flushEnabled'] = '1'
-        if parallel_db_and_view_compaction:
-            payload['parallelDBAndViewCompaction'] = parallel_db_and_view_compaction
-            payload['proxyPort'] = proxy_port
-        payload['ramQuotaMB'] = ram_quota_mb
-        if replica_index:
-            payload['replicaIndex'] = replica_index
-        if replica_number:
-            payload['replicaNumber'] = replica_number
-        if sasl_password:
-            payload['saslPassword'] = sasl_password
-        response = requests.post('http://%s:%s/pools/default/buckets' %(self.host, self.port), auth=(self.username, self.password), data=payload)
-        if response.status_code != 202:
-            # 400 means the document is already created
-            log.error('Unable to create bucket %s on %')
-            raise BadRequest ('Couchbase returned error - status code:%d bucket name:%s - error_string from Couchbase: %s' %(response.status_code, name, response.content))
-        sleep(3)
 
 
-    def delete_datastore(self, datastore_name=''):
-        datastore_name = datastore_name or self.datastore_name
-        log.info('Deleting data store %s' % datastore_name)
-        try:
-            if datastore_name in self._datastore_cache:
-                del self._datastore_cache[datastore_name]
-
-            self.server.delete(datastore_name)
-        except BucketUnavailableException as e:
-            raise NotFound('Couchbase unable to delete bucket named %s on %s. Exception: %s ' %(datastore_name, e.parameters.get('host', None), e._message))
-        # This was added due to Couchbase generating a JSON exception error when trying to delete non-existent bucket.
-        except:
-            raise ServerError('Couchbase returned unknown error')
-
-
-    def list_datastores(self):
-        dbs = [db.name for db in self.server]
-        log.debug('Data stores: %s', str(dbs[0]))
-        return dbs
-
-    def info_datastore(self, datastore_name=""):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        log.debug('Listing information about data store %s', datastore_name)
-        ## Review
-        ## TODO ds.stats??
-        info = ds.stats
-        log.debug('\n\n\nData store info: %s\n\n\n', str(info))
-        return info
-
-    def datastore_exists(self, datastore_name=""):
-        ## TODO Review
-        rest = RestHelper(self.server._rest())
-        return rest.bucket_exists(datastore_name)
-        '''
-        for db in self.server:
-            if db == datastore_name:
-                return True
-        return False
-        '''
-
-    def _list_objects (self, ds, filter=None):
-        view = ds.view(self._get_viewname("association", "by_doc"), include_docs=False, stale="false" )
-        row_ids = [row['id'] for row in view]
-        return row_ids
-
-    ## Todo review
-    ## Right now it only returns ids
-    def list_objects(self, datastore_name=""):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        log.warning('Listing all objects in data store %s' % datastore_name)
-        #objs = [obj for obj in ds]
-        objs = self._list_objects(ds)
-        log.debug('Objects: %s', str(objs))
-        return objs
-
-    def list_object_revisions(self, object_id, datastore_name=""):
-        raise NotImplemented ("List object revisions is not implemented")
-
+    # -------------------------------------------------------------------------
+    # Couch document operations
 
     def create(self, obj, object_id=None, attachments=None, datastore_name=""):
         """
@@ -218,641 +61,67 @@ class CouchbasePyonDataStore(DataStore):
         Returns an identifier and revision number of the object
         """
         if not isinstance(obj, IonObjectBase):
-            raise BadRequest("Obj param is not instance of IonObjectBase: The type is:" + str(type(obj)))
+            raise BadRequest("Obj param is not instance of IonObjectBase")
 
         return self.create_doc(self._ion_object_to_persistence_dict(obj),
                                object_id=object_id, datastore_name=datastore_name,
                                attachments=attachments)
 
-    def create_doc(self, doc, object_id=None, attachments=None, datastore_name=""):
-        """
-        Persists the document using the optionally suggested doc_id, and creates attachments to it.
-        Returns the identifier and version number of the document
-        """
-        ds, datastore_name = self._get_datastore(datastore_name)
-        if '_id' in doc:
-            raise BadRequest("Doc must not have '_id'")
-        if '_rev' in doc:
-            raise BadRequest("Doc must not have '_rev'")
-
-        # Assign an id to doc (recommended in CouchDB documentation)
-        ##doc["_id"] = object_id or self.get_unique_id()
-        doc_id =  object_id or self.get_unique_id()
-        log.debug('Creating new object %s/%s' % (datastore_name, doc_id))
-        log.debug('create doc contents: %s', doc)
-
-        try:
-            ###
-            doc['_id'] = doc_id
-            if isinstance(doc, dict):
-                doc = json.dumps(doc)
-                #TODO should this be add or set?
-            ###opaque, cas, msg = ds.add(doc_id, 0, 0, doc)
-            opaque, cas, msg = ds.set(doc_id, 0, 0, doc)
-            self._count(create=1)
-        except MemcachedError as e:
-            raise NotFound('Object could not be created. Id: %s - Datastore: %s - Exception: %s' % (doc_id, datastore_name, e))
-        except MemcachedTimeoutException as e:
-            raise NotFound('create_doc: Couchbase server timeout:\nId: %s - Datastore: %s - \nException: %s \ndoc: %s' % (doc_id, datastore_name, e, doc))
-
-        # Add the attachments if indicated
-        if attachments is not None:
-            for att_name, att_value in attachments.iteritems():
-                self.create_attachment(doc_id, att_name, att_value['data'], content_type=att_value.get('content_type', ''), datastore_name=datastore_name)
-        log.debug('Create result: %s', str(doc_id))
-        obj_id, version = doc_id, cas
-
-        return obj_id, version
-
-    def create_mult(self, objects, object_ids=None, allow_ids=False):
+    def create_mult(self, objects, object_ids=None):
         if any([not isinstance(obj, IonObjectBase) for obj in objects]):
             raise BadRequest("Obj param is not instance of IonObjectBase")
-        return self.create_doc_mult([self._ion_object_to_persistence_dict(obj) for obj in objects],
-                                    object_ids, allow_ids=allow_ids)
 
-    def create_doc_mult(self, docs, object_ids=None, allow_ids=False):
-        if not allow_ids:
-            if any(["_id" in doc for doc in docs]):
-                raise BadRequest("Docs must not have '_id'")
-            if any(["_rev" in doc for doc in docs]):
-                raise BadRequest("Docs must not have '_rev'")
-        if object_ids and len(object_ids) != len(docs):
-            raise BadRequest("Invalid object_ids")
-        if type(docs) is not list:
-            raise BadRequest("Invalid type for docs:%s" % type(docs))
+        return self.create_doc_mult([self._ion_object_to_persistence_dict(obj) for obj in objects], object_ids)
 
-        if object_ids:
-            for doc, oid in zip(docs, object_ids):
-                doc["_id"] = oid
-        else:
-            #object_ids = [doc.get("_id", None) or self.get_unique_id() for doc in docs]
-            for doc in docs:
-                doc["_id"] = doc.get("_id", None) or self.get_unique_id()
-        '''
-        ### TODO
-        if not object_ids:
-            object_ids = []
-            for doc in docs:
-                doc_id = doc.get("_id", None)
-                if not doc_id:
-                    doc_id = self.get_unique_id()
-                else:
-                    del doc['_id']
-                object_ids.append(doc_id)
-        '''
+    def update(self, obj, datastore_name=""):
+        if not isinstance(obj, IonObjectBase):
+            raise BadRequest("Obj param is not instance of IonObjectBase")
 
+        return self.update_doc(self._ion_object_to_persistence_dict(obj))
 
-        ds, _ = self._get_datastore()
-        #res = db.update(docs)
-        ## Todo. Find other way to do bulk insert
-        ##             opaque, cas, msg = ds.add(doc_id, 0, 0, doc)
-        ###res = [(True, id, cas) for opaque, cas, msg, id in [(ds.set(oid,0,0,doc) + (oid,)) for doc, oid in zip(docs, object_ids)]]
-        res = [(True, id, cas) for opaque, cas, msg, id in [(ds.set(doc['_id'],0,0,doc) + (doc['_id'],)) for doc in docs]]
+    def update_mult(self, objects):
+        if any([not isinstance(obj, IonObjectBase) for obj in objects]):
+            raise BadRequest("Obj param is not instance of IonObjectBase")
 
-        self._count(create_mult_call=1, create_mult_obj=len(docs))
-        if not all([success for success, oid, rev in res]):
-            errors = ["%s:%s" % (oid, rev) for success, oid, rev in res if not success]
-            log.error('create_doc_mult had errors. Successful: %s, Errors: %s'
-                      % (len(res) - len(errors), "\n".join(errors)))
-        else:
-            log.debug('create_doc_mult result: %s', str(res))
+        return self.update_doc_mult([self._ion_object_to_persistence_dict(obj) for obj in objects])
 
-        return res
-
-
-    def create_attachment(self, doc, attachment_name, data, content_type=None, datastore_name=""):
-        """
-        Assumes that the document already exists and creates attachment to it.
-        @param doc can be either id or a document
-        """
-        if not isinstance(attachment_name, str):
-            raise BadRequest("attachment name is not string")
-        if not isinstance(data, str) and not isinstance(data, file):
-            raise BadRequest("data to create attachment is not a str or file")
-
-        ds, _ = self._get_datastore(datastore_name)
-        # If doc is string, assume it is document id
-        doc_id = doc if isinstance(doc, str) else doc.get('_id',None)
-        if not doc_id:
-            raise NotFound("document id is not found in the document for attachment")
-
-        # Update the document to include the attachment id
-        try:
-            obj = self.read_doc(doc_id=doc_id,datastore_name=datastore_name)
-        except:
-            raise NotFound("Document could not found for attachment")
-            # Make sure the attachment name is not already in the document
-        attachment_id = self._get_attachment_id(doc, attachment_name, datastore_name)
-        if attachment_id:
-            raise BadRequest("Attachment name already in use: %s" % attachment_name)
-
-        attachment_id = self._put_attachment(doc=doc, content=data, filename=attachment_name, content_type=content_type)
-
-        if not obj.get(self.__attachment_string, False):
-            obj[self.__attachment_string] = [[attachment_id, attachment_name]]
-        else:
-            obj[self.__attachment_string] += [[attachment_id, attachment_name]]
-
-        # Update the document
-        self.update_doc(obj)
-        #TODO: if update fails, rollback and delete the attachment. Otherwise, there will be attachment without a parent document
-
-        self._count(create_attachment=1)
-        return attachment_id
-
-    def _put_attachment(self, doc, content, filename, content_type, datastore_name=None):
-        # Currently, doc, filename and content type is not used. Keep it for future use
-        content_id =  self.get_unique_id()
-        ds, _ = self._get_datastore(datastore_name)
-        try:
-            #content = json.dumps(content)
-            ds.add(content_id, 0, 0, content)
-        except MemcachedError as e:
-            raise NotFound('Attachment could not be created. Id: %s - Exception: %s' % (content_id, e))
-
-        return content_id
 
     def read(self, object_id, rev_id="", datastore_name=""):
         if not isinstance(object_id, str):
-            raise BadRequest("Object id param is not string. It is %s" % type(object_id) )
-        if rev_id:
-            raise BadRequest("Couchbase doesn't support revision based read")
-        doc = self.read_doc(object_id, rev_id, datastore_name)
+            raise BadRequest("Object id param is not string")
 
-        # Convert doc into Ion object
+        doc = self.read_doc(object_id, rev_id, datastore_name)
         obj = self._persistence_dict_to_ion_object(doc)
-        log.debug('Ion object: %s', str(obj))
 
         return obj
-
-    def read_doc(self, doc_id, rev_id="", datastore_name=""):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        if not rev_id:
-            log.debug('Reading head version of object %s/%s', datastore_name, doc_id)
-            try:
-                status, cas, doc = ds.get(doc_id)
-            except MemcachedError as e:
-                raise NotFound('Object can not be read. ID: %s - Datastore Name: %s. Exception: %s' % (doc_id, datastore_name, e.message))
-            if doc is None:
-                raise NotFound('Object with id %s does not exist.' % str(doc_id))
-            else:
-                ## review
-                doc = json.loads(doc)
-                doc['_rev'] = cas
-                doc['_id'] = str(doc['_id']) #doc_id
-        else:
-            raise NotImplemented ("Read with rev_id is not supported")
-        log.debug('read doc contents: %s', doc)
-        self._count(read=1)
-        return doc
 
     def read_mult(self, object_ids, datastore_name=""):
         if any([not isinstance(object_id, str) for object_id in object_ids]):
             raise BadRequest("Object ids are not string: %s" % str(object_ids))
-            #docs = self.read_doc_mult(object_ids, datastore_name)
-        # Convert docs into Ion objects
-        return self.read_doc_mult(object_ids, datastore_name=datastore_name)
 
-    def read_doc_mult(self, object_ids, datastore_name=""):
-        ## TODO: find other way to do bulk read
-        if not object_ids:
-            return []
-        data, not_found_list = self._read_doc_mult(object_ids,datastore_name)
-        if not_found_list:
-            raise NotFound("\n".join(not_found_list))
+        docs = self.read_doc_mult(object_ids, datastore_name)
+        obj_list = [self._persistence_dict_to_ion_object(doc) for doc in docs]
 
-        return data
-
-    def _read_doc_mult(self, object_ids, datastore_name=""):
-        datastore_name = datastore_name or self.datastore_name
-        data = json.dumps({"keys": object_ids})
-        #todo remove port number
-        response = requests.post('http://%s:%s/%s/_all_docs?include_docs=true&stale=false' %(self.host, '8092', datastore_name), auth=(self.username, self.password), data=data)
-        obj = []
-        notfound_list = []
-        if response.status_code == 404:
-            uri = 'http://%s:%s/%s/_all_docs?include_docs=true' % (self.host, self.port, datastore_name)
-            data = str(data)
-            print ("Couchbase server returned an error. \n Returned code: %d \n Returned content:%s \n URI: %s \n Data:%s " %(response.status_code, response.content, uri, data))
-            log.error("Couchbase server returned an error. \n Returned code: %d \n Returned content:%s \n URI: %s \n Data:%s " %(response.status_code, response.content, uri, data))
-            notfound_list.append('Object id does not exist.')
-        elif response.status_code == 200:
-            rows = json.loads(response.content)
-            for row in rows['rows']:
-                if 'doc' in row and 'json' in row['doc']:
-                    obj.append(self._persistence_dict_to_ion_object(row['doc']['json']))
-                elif 'error' in row:
-                    obj.append(None)
-                    notfound_list.append('Object with id %s does not exist.' % row['key'])
-        else:
-            uri = 'http://%s:%s/%s/_all_docs?include_docs=true' % (self.host, self.port, datastore_name)
-            data = str(data)
-            raise ServerError("Couchbase server returned an error. \n Returned code: %d \n Returned content:%s \n URI: %s \n Data:%s " %(response.status_code, response.content, uri, data))
-
-        return obj, notfound_list
+        return obj_list
 
 
-    def read_attachment(self, doc, attachment_name, datastore_name=""):
-        if not isinstance(attachment_name, str):
-            raise BadRequest("Attachment_name param is not str")
-        ds, datastore_name = datastore_name or self._get_datastore(datastore_name)
-
-        # Get attachment id from the document
-        attachment_id = self._get_attachment_id(doc, attachment_name, datastore_name)
-        if not attachment_id:
-            raise NotFound ("Document doesn't have attachment")
-
-        log.debug('Fetching attachment %s of document %s/%s', attachment_name, datastore_name, doc)
-        attachment = self._get_attachment(attachment_id, attachment_name, datastore_name)
-
-        if attachment is None:
-            raise NotFound('Attachment %s does not exist in document %s.%s.', attachment_name, datastore_name, doc)
-        else:
-            log.debug('Reading attachment content: %s', attachment)
-
-        if not isinstance(attachment, str):
-            raise NotFound('Attachment read is not a string')
-
-        log.debug('Read content of attachment: %s of document %s/%s', attachment_name, datastore_name, doc)
-
-        self._count(read_attachment=1)
-        return attachment
-
-    def _get_attachment (self, doc_id, attachment_name, datastore_name=""):
-        if not doc_id:
-            raise BadRequest('Attachment id could not be found')
-        ds, _  = self._get_datastore(datastore_name)
-        try:
-            status, cas, attachment = ds.get(doc_id)
-        except MemcachedError as e:
-            raise NotFound('Object with id %s could not be read. Exception: %s' % (doc_id, e.message))
-        if attachment is None:
-            raise NotFound('Object with id %s does not exist.' % str(doc_id))
-
-        return attachment
-
-    def _get_attachment_id (self, doc, attachment_name, datastore_name):
-        doc_id = doc if isinstance(doc, str) else doc.get('_id',None)
-        if not doc_id:
-            raise BadRequest("document id is not found in the document ")
-            # Gets attachment id from the document
-        try:
-            obj = self.read_doc(doc_id, datastore_name=datastore_name)
-        except NotFound:
-            raise NotFound ('Document could not be found for attachment read')
-
-        if not obj.get(self.__attachment_string, False):
-            attachment_id = None
-        else:
-            # Find the attachment id from the list
-            #   "__attachments": [
-            #    [
-            #        "attachment id"
-            #        "attachment name"
-            #    ],
-            #    [
-            #        "attachment id"
-            #        "attachment name"
-            #    ]
-            #  ],
-            attachment_id = [attachment_meta_data[0] for attachment_meta_data in obj[self.__attachment_string] if attachment_meta_data[1] == attachment_name]
-            attachment_id = attachment_id[0] if attachment_id else None
-
-        return attachment_id
-
-
-    #Todo: should this have datastore_name??
-    def list_attachments(self, doc):
-        """
-        Returns the a list of attachments for the document, as a dict of dicts, key'ed by name with
-        nested keys 'data' for the content and 'content-type'.
-        @param doc  accepts either str (meaning an id) or dict (meaning a full document).
-        """
-        attachment_metadata = self._list_attachment_metadata(doc)
-        attachments = None
-        if attachment_metadata:
-            # Todo: Currently content type is always empty
-            attachments =[dict({'data' : self._get_attachment(attachment[0], attachment_name=attachment[1]), 'attachment_name':attachment[1], 'content-type':'' }) for attachment in attachment_metadata]
-
-        return attachments
-
-    def _list_attachment_metadata (self, doc):
-        if isinstance(doc, dict):
-            obj = self.read_doc(doc_id=doc["_id"])
-        elif isinstance(doc, str):
-            obj = self.read_doc(doc_id=doc)
-        else:
-            raise BadRequest("document id is not found in the document ")
-
-        if not obj.get(self.__attachment_string, False):
-            attachment_metadata = None
-        else:
-            attachment_metadata = [attachment_metadata for attachment_metadata in obj[self.__attachment_string]]
-
-        return attachment_metadata
-
-    def update(self, obj, datastore_name=""):
-        if not isinstance(obj, IonObjectBase):
-            raise BadRequest("Obj param is not instance of IonObjectBase. The type is: " + str(type(obj)))
-        return self.update_doc(self._ion_object_to_persistence_dict(obj))
-
-    def update_doc(self, doc, datastore_name=""):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        if '_id' not in doc:
-            raise BadRequest("Doc must have '_id'")
-            ## Review
-        #if '_rev' not in doc:
-        #    raise BadRequest("Doc must have '_rev'")
-
-        log.debug('update doc contents: %s', doc)
-        try:
-            key = str(doc['_id'])
-            if isinstance(doc, dict):
-                doc = json.dumps(doc)
-            opaque, cas, msg = ds.replace(key, 0, 0, doc)
-            self._count(update=1)
-        except MemcachedError as e:
-            raise NotFound('Object with id %s could not be updated. Exception: %s' % (key, e))
-        log.debug('Update result: %s', str(key))
-        return key, cas
-
-    def update_attachment(self, doc, attachment_name, data, content_type=None, datastore_name=""):
-        log.debug("updating attachment %s", attachment_name)
-        attachment_id = self._get_attachment_id(doc, attachment_name, datastore_name)
-        if not attachment_id:
-            # if attachment not found, create it
-            self.create_attachment(doc=doc, data=data, attachment_name=attachment_name, content_type=content_type, datastore_name=datastore_name)
-        else:
-            ds, _ = self._get_datastore(datastore_name)
-            try:
-                _, _, _ = ds.replace(attachment_id, 0, 0, data)
-            except MemcachedError as e:
-                raise NotFound('Attachment with with document id %s could not be updated. Exception: %s' % (attachment_id, e))
-        log.debug("updated attachment %s", attachment_name)
-
-    def delete(self, obj, datastore_name="", del_associations=False):
+    def delete(self, obj, datastore_name=""):
         if not isinstance(obj, IonObjectBase) and not isinstance(obj, str):
             raise BadRequest("Obj param is not instance of IonObjectBase or string id")
         if type(obj) is str:
-            self.delete_doc(obj, datastore_name=datastore_name, del_associations=del_associations)
+            self.delete_doc(obj, datastore_name=datastore_name)
         else:
             if '_id' not in obj:
                 raise BadRequest("Doc must have '_id'")
-                #if '_rev' not in obj:
-            #    raise BadRequest("Doc must have '_rev'")
-            self.delete_doc(self._ion_object_to_persistence_dict(obj), datastore_name=datastore_name, del_associations=del_associations)
-
-    def delete_doc(self, doc, datastore_name="", del_associations=False):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        doc_id = doc if type(doc) is str else doc["_id"]
-        log.debug('Deleting object %s/%s', datastore_name, doc_id)
-
-        if del_associations:
-            assoc_ids = self.find_associations(anyside=doc_id, id_only=True)
-            self.delete_doc_mult(assoc_ids)
-        #            for aid in assoc_ids:
-        #                self.delete(aid, datastore_name=datastore_name)
-        #            log.info("Deleted %n associations for object %s", len(assoc_ids), doc_id)
-
-        elif self._is_in_association(doc_id, datastore_name):
-            bad_doc = self.read(doc_id)
-            if doc:
-                log.warn("XXXXXXX Attempt to delete %s object %s that still has associations" % (bad_doc.type_, doc_id))
-            else:
-                log.warn("XXXXXXX Attempt to delete object %s that still has associations" % doc_id)
-                #           raise BadRequest("Object cannot be deleted until associations are broken")
-
-        try:
-            ds.delete(doc_id)
-            self._count(delete=1)
-        except MemcachedError as e:
-            raise NotFound('Object with id %s could not be deleted. Exception: %s' % (doc_id, e))
+            self.delete_doc(self._ion_object_to_persistence_dict(obj),
+                            datastore_name=datastore_name)
 
     def delete_mult(self, object_ids, datastore_name=None):
         return self.delete_doc_mult(object_ids, datastore_name)
 
-    def delete_doc_mult(self, object_ids, datastore_name=None):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        # Todo find other way to do bulk delete
-        return [ds.delete(id) for id in object_ids]
 
-    def delete_attachment(self, doc, attachment_name, datastore_name=""):
-        """
-        Deletes an attachment from a document.
-        """
-        if not isinstance(attachment_name, str):
-            raise BadRequest("attachment_name is not a string")
-        doc_id = doc if isinstance(doc, str) else doc.get('_id', None)
-
-        attachment_id = self._get_attachment_id(doc_id, attachment_name, datastore_name)
-        if not attachment_id:
-            raise NotFound('Attachment could not be found for a document id: %s' % doc_id)
-
-        ds, datastore_name = self._get_datastore(datastore_name)
-
-        log.debug('Deleting attachment of document %s/%s' %(datastore_name, doc["_id"]))
-        self.delete_doc(str(attachment_id))
-        self._delete_attachment_meta_data(doc_id, attachment_name, datastore_name)
-        log.debug('Deleted attachment: %s', attachment_name)
-        self._count(delete_attachment=1)
-
-    def _delete_attachment_meta_data(self, doc_id,attachment_name, datastore_name=""):
-        """
-        Removes attachment id and name from the document
-        Assumes attachment id and name already in the document
-        """
-        if not type(doc_id) is str:
-            raise BadRequest('doc_id must be string')
-            # Get the document
-        doc = self.read_doc(doc_id, datastore_name=datastore_name)
-        # Remove attachment id and name from the document
-        attachment_meta_data = [attachment_meta_data for attachment_meta_data in doc[self.__attachment_string] if attachment_meta_data[1] != attachment_name]
-        doc[self.__attachment_string] = attachment_meta_data
-        self.update_doc(doc, datastore_name=datastore_name)
-
-
-    def create_association(self, subject=None, predicate=None, obj=None, assoc_type=None):
-        """
-        Create an association between two IonObjects with a given predicate
-        """
-        #if assoc_type:
-        #if assoc_type:
-        #    raise BadRequest("assoc_type deprecated")
-        if not (subject and predicate and obj):
-            raise BadRequest("Association must have all elements set")
-        if type(subject) is str:
-            subject_id = subject
-            subject = self.read(subject_id)
-            subject_type = subject._get_type()
-        else:
-            #if "_id" not in subject or "_rev" not in subject:
-            if "_id" not in subject:
-                raise BadRequest("Subject id or rev not available")
-            subject_id = subject._id
-            subject_type = subject._get_type()
-
-        if type(obj) is unicode:
-            obj = str(obj)
-        if type(obj) is str:
-            object_id = obj
-            obj = self.read(object_id)
-            object_type = obj._get_type()
-        else:
-            #if "_id" not in obj or "_rev" not in obj:
-            if "_id" not in obj:
-                raise BadRequest("Object id or rev not available")
-            object_id = obj._id
-            object_type = obj._get_type()
-
-        # Check that subject and object type are permitted by association definition
-        # Note: Need import here, so that import orders are not screwed up
-        from pyon.core.registry import getextends
-        from pyon.ion.resource import Predicates
-        from pyon.core.bootstrap import IonObject
-
-        try:
-            pt = Predicates.get(predicate)
-        except AttributeError:
-            raise BadRequest("Predicate unknown %s" % predicate)
-        if not subject_type in pt['domain']:
-            found_st = False
-            for domt in pt['domain']:
-                if subject_type in getextends(domt):
-                    found_st = True
-                    break
-            if not found_st:
-                raise BadRequest("Illegal subject type %s for predicate %s" % (subject_type, predicate))
-        if not object_type in pt['range']:
-            found_ot = False
-            for rant in pt['range']:
-                if object_type in getextends(rant):
-                    found_ot = True
-                    break
-            if not found_ot:
-                raise BadRequest("Illegal object type %s for predicate %s" % (object_type, predicate))
-
-        # Finally, ensure this isn't a duplicate
-        assoc_list = self.find_associations(subject, predicate, obj, id_only=False)
-        if len(assoc_list) != 0:
-            assoc = assoc_list[0]
-            raise BadRequest("Association between %s and %s with predicate %s already exists" % (subject, obj, predicate))
-
-        assoc = IonObject("Association",
-                          s=subject_id, st=subject_type,
-                          p=predicate,
-                          o=object_id, ot=object_type,
-                          ts=get_ion_ts())
-        self._count(_create_assoc=1)
-        return self.create(assoc, create_unique_association_id())
-
-    def delete_association(self, association=''):
-        """
-        Delete an association between two IonObjects
-        @param association  Association object, association id or 3-list of [subject, predicate, object]
-        """
-        if type(association) in (list, tuple) and len(association) == 3:
-            subject, predicate, obj = association
-            assoc_id_list = self.find_associations(subject=subject, predicate=predicate, obj=obj, id_only=True)
-            success = True
-            for aid in assoc_id_list:
-                success = success and self.delete(aid)
-            self._count(_delete_assoc=1)
-            return success
-        else:
-            self._count(_delete_assoc=1)
-            return self.delete(association)
-
-    def _get_viewname(self, design, name):
-        return "_design/%s/_view/%s" % (design, name)
-
-    def _define_views(self, datastore_name=None, profile=None, keepviews=False):
-        datastore_name = datastore_name or self.datastore_name
-        profile = profile or self.profile
-        log.debug('Define views datastore: %s profile: %s' %(datastore_name,profile))
-
-        ds_views = get_couchdb_view_designs(profile)
-        for design, viewdef in ds_views.iteritems():
-            self._define_view(design, viewdef, datastore_name=datastore_name, keepviews=keepviews)
-
-    def _get_view_names(self,design_name,  datastore_name=None) :
-        ds, datastore_name = self._get_datastore(datastore_name)
-        design_path = "_design/%s" % design_name
-        design_doc = ds[design_path]
-        return [view.name for view in design_doc.views()]
-
-
-    def _define_view(self, design, viewdef, datastore_name=None, keepviews=False):
-        ds, datastore_name = self._get_datastore(datastore_name)
-        viewname = "_design/%s" % design
-
-        if keepviews and viewname in ds.design_docs():
-        #if keepviews and viewname in self._get_view_names(design, datastore_name):
-            return
-        try:
-            ##TODO review this
-            ### del ds[viewname]
-            rest = self.server._rest()
-            rest.delete_design_doc(bucket=datastore_name, design_doc=design)
-        except:
-            pass
-        ds[viewname] = dict(views=viewdef)
-
-    def _update_views(self, datastore_name="", profile=None):
-        ds, datastore_name = self._get_datastore(datastore_name)
-
-        profile = profile or self.profile
-        ds_views = get_couchdb_view_designs(profile)
-
-        for design, viewdef in ds_views.iteritems():
-            for viewname in viewdef:
-                try:
-                    rows = ds.view("_design/%s/_view/%s" % (design, viewname))
-                    log.debug("View %s/_design/%s/_view/%s: %s rows", datastore_name, design, viewname, len(rows))
-                except Exception, ex:
-                    log.exception("Problem with view %s/_design/%s/_view/%s", datastore_name, design, viewname)
-
-    _refresh_views = _update_views
-
-    def _delete_views(self, datastore_name="", profile=None):
-        ds, datastore_name = self._get_datastore(datastore_name)
-
-        profile = profile or self.profile
-        ds_views = get_couchdb_view_designs(profile)
-
-        for design, viewdef in ds_views.iteritems():
-            try:
-                #TODO: review
-                ###del ds["_design/%s" % design]
-                rest = self.server._rest()
-                rest.delete_design_doc(bucket=datastore_name, design_doc=design)
-            except ResourceNotFound:
-                pass
-
-    def _get_view_args(self, all_args):
-        """
-        @brief From given all_args dict, extract all entries that are valid CouchDB view options.
-        @see http://wiki.apache.org/couchdb/HTTP_view_API
-        """
-        view_args = dict((k, v) for k, v in all_args.iteritems() if k in ('descending', 'stale', 'skip', 'inclusive_end', 'update_seq') and v is not None)
-        limit = int(all_args.get('limit', 0)) if all_args.get('limit', None) is not None else 0
-        if limit > 0:
-            view_args['limit'] = limit
-        return view_args
-
-
-    def _is_in_association(self, obj_id, datastore_name=""):
-        log.debug("_is_in_association(%s)", obj_id)
-        if not obj_id:
-            raise BadRequest("Must provide object id")
-
-        assoc_ids = self.find_associations(anyside=obj_id, id_only=True, limit=1)
-        if assoc_ids:
-            log.debug("Object found as object in associations: %s", assoc_ids)
-            return True
-
-        return False
-
-
+    # -------------------------------------------------------------------------
+    # View operations
 
     def find_objects_mult(self, subjects, id_only=False):
         """
@@ -861,7 +130,7 @@ class CouchbasePyonDataStore(DataStore):
         ds, datastore_name = self._get_datastore()
         validate_is_instance(subjects, list, 'subjects is not a list of resource_ids')
         view_args = dict(keys=subjects, include_docs=True)
-        results = self.query_view(self._get_viewname("association", "by_bulk"), view_args)
+        results = self.query_view(self._get_view_name("association", "by_bulk"), view_args)
         ids = [i['value'] for i in results]
         assocs = [i['doc'] for i in results]
         self._count(find_assocs_mult_call=1, find_assocs_mult_obj=len(ids))
@@ -877,7 +146,7 @@ class CouchbasePyonDataStore(DataStore):
         ds, datastore_name = self._get_datastore()
         validate_is_instance(objects, list, 'objects is not a list of resource_ids')
         view_args = dict(keys=objects, include_docs=True)
-        results = self.query_view(self._get_viewname("association", "by_subject_bulk"), view_args)
+        results = self.query_view(self._get_view_name("association", "by_subject_bulk"), view_args)
         ids = [i['value'] for i in results]
         assocs = [i['doc'] for i in results]
         self._count(find_assocs_mult_call=1, find_assocs_mult_obj=len(ids))
@@ -895,7 +164,7 @@ class CouchbasePyonDataStore(DataStore):
         if not subject:
             raise BadRequest("Must provide subject")
         if object_type and not predicate:
-            raise BadRequest("Cannot provide object type without a predictate")
+            raise BadRequest("Cannot provide object type without a predicate")
 
         ds, datastore_name = self._get_datastore()
 
@@ -908,7 +177,7 @@ class CouchbasePyonDataStore(DataStore):
                 subject_id = subject._id
 
         view_args = self._get_view_args(kwargs)
-        #view = ds.view(self._get_viewname("association", "by_sub"), **view_args)
+        #view = ds.view(self._get_view_name("association", "by_sub"), **view_args)
         key = [subject_id]
         if predicate:
             key.append(predicate)
@@ -916,11 +185,11 @@ class CouchbasePyonDataStore(DataStore):
                 key.append(object_type)
         endkey = self._get_endkey(key)
         #rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("association", "by_sub"), start_key=key, end_key=endkey, stale="false", **view_args)
+        rows = ds.view(self._get_view_name("association", "by_sub"), start_key=key, end_key=endkey, stale="false", **view_args)
 
         '''
         import pprint
-        print "\n\n view objects: ", self._get_viewname("association", "by_sub"), " key:", key, " endkey", endkey
+        print "\n\n view objects: ", self._get_view_name("association", "by_sub"), " key:", key, " endkey", endkey
         print pprint.pformat(rows)
         print "\n\n"
         '''
@@ -966,7 +235,7 @@ class CouchbasePyonDataStore(DataStore):
                 key.append(subject_type)
         endkey = self._get_endkey(key)
         #rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("association", "by_obj"), start_key=key, end_key=endkey, stale="false", **view_args)
+        rows = ds.view(self._get_view_name("association", "by_obj"), start_key=key, end_key=endkey, stale="false", **view_args)
 
         sub_assocs = [self._persistence_dict_to_ion_object(row['value']) for row in rows]
         sub_ids = [str(assoc.s) for assoc in sub_assocs]                                    ## Review str()
@@ -1029,35 +298,35 @@ class CouchbasePyonDataStore(DataStore):
             if predicate:
                 key.append(predicate)
             endkey = self._get_endkey(key)
-            rows = ds.view(self._get_viewname("association", "by_match"), start_key=key, end_key=endkey, stale="false", **view_args)
+            rows = ds.view(self._get_view_name("association", "by_match"), start_key=key, end_key=endkey, stale="false", **view_args)
         elif subject:
             key = [subject_id]
             if predicate:
                 key.append(predicate)
             endkey = self._get_endkey(key)
-            rows = ds.view(self._get_viewname("association", "by_sub"), start_key=key, end_key=endkey, stale="false", **view_args)
+            rows = ds.view(self._get_view_name("association", "by_sub"), start_key=key, end_key=endkey, stale="false", **view_args)
         elif obj:
             key = [object_id]
             if predicate:
                 key.append(predicate)
             endkey = self._get_endkey(key)
-            rows = ds.view(self._get_viewname("association", "by_obj"), start_key=key, end_key=endkey, stale="false", **view_args)
+            rows = ds.view(self._get_view_name("association", "by_obj"), start_key=key, end_key=endkey, stale="false", **view_args)
         elif anyside:
             if predicate:
                 key = [anyside, predicate]
                 endkey = self._get_endkey(key)
-                rows = ds.view(self._get_viewname("association", "by_idpred"), start_key=key, end_key=endkey, stale="false", **view_args)
+                rows = ds.view(self._get_view_name("association", "by_idpred"), start_key=key, end_key=endkey, stale="false", **view_args)
             elif type(anyside_ids[0]) is str:
                 anyside_ids = json.dumps(anyside_ids)
-                rows = ds.view(self._get_viewname("association", "by_id"), keys=anyside_ids, **view_args)
+                rows = ds.view(self._get_view_name("association", "by_id"), keys=anyside_ids, **view_args)
             else:
                 ## Review json dumps
                 anyside_ids = json.dumps(anyside_ids)
-                rows = ds.view(self._get_viewname("association", "by_idpred"), keys=anyside_ids, **view_args)
+                rows = ds.view(self._get_view_name("association", "by_idpred"), keys=anyside_ids, **view_args)
         elif predicate:
             key = [predicate]
             endkey = self._get_endkey(key)
-            rows = ds.view(self._get_viewname("association", "by_pred"), start_key=key, end_key=endkey, stale="false", **view_args)
+            rows = ds.view(self._get_view_name("association", "by_pred"), start_key=key, end_key=endkey, stale="false", **view_args)
         else:
             raise BadRequest("Illegal arguments")
 
@@ -1128,16 +397,16 @@ class CouchbasePyonDataStore(DataStore):
 
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_type"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_type"), include_docs=(not id_only), **filter)
         if restype:
             key = [restype]
             endkey = self._get_endkey(key)
             ##rows = view[key:endkey]
-            rows = ds.view(self._get_viewname("resource", "by_type"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+            rows = ds.view(self._get_view_name("resource", "by_type"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
         else:
             # Returns ALL documents, only limited by filter
             ##rows = view
-            rows = ds.view(self._get_viewname("resource", "by_type"), include_docs=(not id_only), **filter)
+            rows = ds.view(self._get_view_name("resource", "by_type"), include_docs=(not id_only), **filter)
 
         #res_assocs = [dict(type=row['key'][0], lcstate=row['key'][1], name=row['key'][2], id=row['id']) for row in rows]
         res_assocs = [dict(type=row['key'][0], name=row['key'][1], id=row['id']) for row in rows]
@@ -1154,13 +423,13 @@ class CouchbasePyonDataStore(DataStore):
             lcstate,_ = lcstate.split("_", 1)
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_lcstate"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_lcstate"), include_docs=(not id_only), **filter)
         key = [1, lcstate] if lcstate in CommonResourceLifeCycleSM.AVAILABILITY else [0, lcstate]
         if restype:
             key.append(restype)
         endkey = self._get_endkey(key)
         ##rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_lcstate"), include_docs=(not id_only),  start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_lcstate"), include_docs=(not id_only),  start_key=key, end_key=endkey, stale="false", **filter)
 
         res_assocs = [dict(lcstate=row['key'][1], type=row['key'][2], name=row['key'][3], id=row['id']) for row in rows]
 
@@ -1174,13 +443,13 @@ class CouchbasePyonDataStore(DataStore):
             raise BadRequest('id_only must be type bool, not %s' % type(id_only))
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_name"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_name"), include_docs=(not id_only), **filter)
         key = [name]
         if restype:
             key.append(restype)
         endkey = self._get_endkey(key)
         #rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_name"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_name"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
 
         #res_assocs = [dict(name=row['key'][0], type=row['key'][1], lcstate=row['key'][2], id=row['id']) for row in rows]
         res_assocs = [dict(name=row['key'][0], type=row['key'][1], id=row['id']) for row in rows]
@@ -1196,13 +465,13 @@ class CouchbasePyonDataStore(DataStore):
             raise BadRequest('id_only must be type bool, not %s' % type(id_only))
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_keyword"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_keyword"), include_docs=(not id_only), **filter)
         key = [keyword]
         if restype:
             key.append(restype)
         endkey = self._get_endkey(key)
         ##rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_keyword"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_keyword"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
 
         res_assocs = [dict(keyword=row['key'][0], type=row['key'][1], id=row['id']) for row in rows]
         log.debug("find_res_by_keyword() found %s objects", len(res_assocs))
@@ -1217,13 +486,13 @@ class CouchbasePyonDataStore(DataStore):
             raise BadRequest('id_only must be type bool, not %s' % type(id_only))
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_nestedtype"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_nestedtype"), include_docs=(not id_only), **filter)
         key = [nested_type]
         if restype:
             key.append(restype)
         endkey = self._get_endkey(key)
         #rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_nestedtype"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_nestedtype"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
 
         res_assocs = [dict(nested_type=row['key'][0], type=row['key'][1], id=row['id']) for row in rows]
         log.debug("find_res_by_nested_type() found %s objects", len(res_assocs))
@@ -1238,13 +507,13 @@ class CouchbasePyonDataStore(DataStore):
             raise BadRequest('id_only must be type bool, not %s' % type(id_only))
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_attribute"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_attribute"), include_docs=(not id_only), **filter)
         key = [restype, attr_name]
         if attr_value:
             key.append(attr_value)
         endkey = self._get_endkey(key)
         ##rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_attribute"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_attribute"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
 
         res_assocs = [dict(type=row['key'][0], attr_name=row['key'][1], attr_value=row['key'][2], id=row['id']) for row in rows]
         log.debug("find_res_by_attribute() found %s objects", len(res_assocs))
@@ -1261,7 +530,7 @@ class CouchbasePyonDataStore(DataStore):
             raise BadRequest('id_only must be type bool, not %s' % type(id_only))
         filter = filter if filter is not None else {}
         ds, datastore_name = self._get_datastore()
-        ##view = ds.view(self._get_viewname("resource", "by_altid"), include_docs=(not id_only), **filter)
+        ##view = ds.view(self._get_view_name("resource", "by_altid"), include_docs=(not id_only), **filter)
         key = []
         if alt_id:
             key.append(alt_id)
@@ -1270,7 +539,7 @@ class CouchbasePyonDataStore(DataStore):
 
         endkey = self._get_endkey(key)
         ##rows = view[key:endkey]
-        rows = ds.view(self._get_viewname("resource", "by_altid"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
+        rows = ds.view(self._get_view_name("resource", "by_altid"), include_docs=(not id_only), start_key=key, end_key=endkey, stale="false", **filter)
 
         if alt_id_ns and not alt_id:
             res_assocs = [dict(alt_id=row['key'][0], alt_id_ns=row['key'][1], id=row['id']) for row in rows if row['key'][1] == alt_id_ns]
@@ -1295,144 +564,36 @@ class CouchbasePyonDataStore(DataStore):
         # TODO: Refactor common code out of above find functions
         pass
 
-    ## Need to do more tests
     def find_by_view(self, design_name, view_name, key=None, keys=None, start_key=None, end_key=None,
-                     id_only=True, convert_doc=True, **kwargs):
+                     id_only=True, convert_doc=True, convert_value=True, **kwargs):
         """
-        @brief Generic find function using an defined index
-        @retval Returns a list of triples: (object _id, index key, Document/object or None)
+        Generic find function using a defined index
+        @param design_name  design document
+        @param view_name  view name
+        @param key  specific key to find
+        @param keys  list of keys to find
+        @param start_key  find range start value
+        @param end_key  find range end value
+        @param id_only  if True, the 4th element of each triple is the document
+        @param convert_doc  if True, make IonObject out of doc
+        @param convert_value  if True, make IonObject out of value
+        @retval Returns a list of 4-tuples: (document id, index key, index value, document)
         """
-        if type(id_only) is not bool:
-            raise BadRequest('id_only must be type bool, not %s' % type(id_only))
-        ds, datastore_name = self._get_datastore()
+        res_rows = self.find_docs_by_view(design_name=design_name, view_name=view_name, key=key, keys=keys,
+                                          start_key=start_key, end_key=end_key, id_only=id_only, **kwargs)
 
-        view_args = self._get_view_args(kwargs)
-        view_args['include_docs'] = (not id_only)
-        view_doc = design_name if design_name == "_all_docs" else self._get_viewname(design_name, view_name)
-        ##if keys:
-        ##    view_args['keys'] = keys
-        #@view = ds.view(view_doc, **view_args)
-        if key is not None:
-            rows = ds.view(view_doc, key=key,  **view_args)
-            log.info("find_by_view(%s): key=%s", view_doc, key)
-        elif keys:
-            #rows = view
-            rows = ds.view(view_doc, keys=keys,  **view_args)
-            log.info("find_by_view(%s): keys=%s", view_doc, str(keys))
-        elif start_key and end_key:
-            startkey = start_key or []
-            endkey = list(end_key) or []
-            endkey.append(END_MARKER)
-            log.info("find_by_view(%s): start_key=%s to end_key=%s", view_doc, startkey, endkey)
-            if view_args.get('descending', False):
-                ##rows = view[endkey:startkey]
-                rows = ds.view(view_doc,  start_key=endkey, end_key=startkey, stale="false", **view_args)
-            else:
-                ##rows = view[startkey:endkey]
-                rows = ds.view(view_doc,  start_key=key, end_key=endkey, stale="false", **view_args)
-        else:
-            #rows = view
-            rows = ds.view(view_doc, **view_args)
+        res_rows = [(rid, key,
+                     self._persistence_dict_to_ion_object(value) if convert_value and isinstance(value, dict) else value,
+                     self._persistence_dict_to_ion_object(doc) if convert_doc and isinstance(doc, dict) else doc)
+                    for rid,key,value,doc in res_rows]
 
-        ###rows = [self._persistence_dict_to_ion_object(dict(_id=row['id'], **row)) for row in rows]
-        rows = [self._persistence_dict_to_ion_object(row) for row in rows]
-
-        if id_only:
-            if convert_doc:
-                res_rows = [(row['id'], row['key'], self._persistence_dict_to_ion_object(row['value'])) for row in rows]
-            else:
-                res_rows = [(row['id'], row['key'], row['value']) for row in rows]
-        else:
-            if convert_doc:
-                res_rows = [(row['id'], row['key'], self._persistence_dict_to_ion_object(row['doc'])) for row in rows]
-            else:
-                res_rows = [(row['id'], row['key'], row['doc']) for row in rows]
-
-        self._count(find_by_view_call=1, find_by_view_obj=len(res_rows))
-
-        log.info("find_by_view() found %s objects" % (len(res_rows)))
+        log.debug("find_by_view() found %s objects" % (len(res_rows)))
         return res_rows
-
-    def _get_endkey(self, startkey):
-        if startkey is None or type(startkey) is not list:
-            raise BadRequest("Cannot create endkey for type %s" % type(startkey))
-        endkey = list(startkey)
-        endkey.append(END_MARKER)
-        return endkey
-
-    def _ion_object_to_persistence_dict(self, ion_object):
-        if ion_object is None: return None
-
-        obj_dict = self._io_serializer.serialize(ion_object)
-        return obj_dict
-
-    def _persistence_dict_to_ion_object(self, obj_dict):
-        if obj_dict is None: return None
-        ion_object = self._io_deserializer.deserialize(obj_dict)
-        return ion_object
-
-    def query_view(self, view_name='', opts={}, datastore_name=''):
-        '''
-        query_view is a straight through method for querying a view in CouchDB. query_view provides us the interface
-        to the view structure in couch, in lieu of implementing a method for every type of query we could want, we
-        now have the capability for clients to make queries to couch in a straight-through manner.
-        '''
-        ds, datastore_name = self._get_datastore(datastore_name)
-
-        # Actually obtain the results and place them in rows
-        rows = ds.view(view_name, **opts)
-
-        # Parse the results and convert the results into ionobjects and python types.
-        result = self._parse_results(rows)
-
-        return result
-
-    def custom_query(self, map_fun, reduce_fun=None, datastore_name='', **options):
-        '''
-        custom_query sets up a temporary view in couchdb, the map_fun is a string consisting
-        of the javascript map function
-
-        Warning: Please note that temporary views are not suitable for use in production,
-        as they are really slow for any database with more than a few dozen documents.
-        You can use a temporary view to experiment with view functions, but switch to a
-        permanent view before using them in an application.
-        '''
-        ds, datastore_name = self._get_datastore(datastore_name)
-        res = ds.query(map_fun, reduce_fun, **options)
-
-        return self._parse_results(res)
-
 
     def _parse_results(self, doc):
         ''' Parses a complex object and organizes it into basic types
         '''
         ret = {}
-
-        #-------------------------------
-        # Handle ViewResults type (CouchDB type)
-        #-------------------------------
-        # \_ Ignore the meta data and parse the rows only
-        if isinstance(doc, ViewResults):
-            try:
-                ret = self._parse_results(doc.rows)
-            except ResourceNotFound as e:
-                raise BadRequest('The desired resource does not exist.')
-
-            return ret
-
-        #-------------------------------
-        # Handle A Row (CouchDB type)
-        #-------------------------------
-        # \_ Split it into a dict with a key and a value
-        #    Recursively parse down through the structure.
-        if isinstance(doc, Row):
-            if 'id' in doc:
-                ret['id'] = doc['id']
-            ret['key'] = self._parse_results(doc['key'])
-            ret['value'] = self._parse_results(doc['value'])
-            if 'doc' in doc:
-                ret['doc'] = self._parse_results(doc['doc'])
-            return ret
 
         #-------------------------------
         # Handling a list
@@ -1464,9 +625,17 @@ class CouchbasePyonDataStore(DataStore):
         return doc
 
 
-    def _count(self, datastore=None, **kwargs):
-        datastore = datastore or self.datastore_name
-        self._stats.count(namespace=datastore, **kwargs)
+    def _ion_object_to_persistence_dict(self, ion_object):
+        if ion_object is None:
+            return None
 
-    def get_unique_id(self):
-        return uuid4().hex
+        obj_dict = self._io_serializer.serialize(ion_object)
+        return obj_dict
+
+    def _persistence_dict_to_ion_object(self, obj_dict):
+        if obj_dict is None:
+            return None
+
+        ion_object = self._io_deserializer.deserialize(obj_dict)
+        return ion_object
+
